# internal/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Connector:
    properties:
      config:
        $ref: '#/definitions/models.ConnectorConfig'
      created_at:
        type: string
      deleted_at:
        type: string
      folder_id:
        type: string
      id:
        type: string
      is_default:
        type: boolean
      max_rows:
        type: integer
      name:
        type: string
      org_id:
        type: string
      table_allowlist:
        items:
          type: string
        type: array
      table_denylist:
        items:
          type: string
        type: array
      timeout_seconds:
        type: integer
      type:
        $ref: '#/definitions/models.ConnectorType'
      updated_at:
        type: string
    type: object
  models.ConnectorConfig:
    properties:
      database:
        type: string
      host:
        type: string
      password:
        type: string
      port:
        type: integer
      ssl_mode:
        type: string
      user:
        type: string
    type: object
  models.ConnectorType:
    enum:
    - postgres
    - clickhouse
    - opensearch
    type: string
    x-enum-varnames:
    - ConnectorPostgres
    - ConnectorClickHouse
    - ConnectorOpenSearch
...
host: localhost:8080
info:
  contact: {}
  description: Aether Notebooks API — collaborative SQL/data notebook platform
  title: Aether API
  version: 1.0.0
paths:
...
  /connectors:
    get:
      description: List all connectors for the current organization
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Connector'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List connectors
      tags:
      - connectors
    post:
      consumes:
      - application/json
      description: Create a new database connector
      parameters:
      - description: Connector details
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Connector'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a connector
      tags:
      - connectors
  /connectors/{id}:
    delete:
      description: Delete a database connector
      parameters:
      - description: Connector ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a connector
      tags:
      - connectors
    get:
      description: Get a single connector by ID
      parameters:
      - description: Connector ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Connector'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get connector
      tags:
      - connectors
    put:
      consumes:
      - application/json
      description: Update a database connector's configuration
      parameters:
      - description: Connector ID
        in: path
        name: id
        required: true
        type: string
      - description: Connector updates
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Connector'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a connector
      tags:
      - connectors
  /connectors/{id}/databases:
    get:
      description: List all databases available through a connector
      parameters:
      - description: Connector ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List connector databases
      tags:
      - connectors
  /connectors/{id}/default:
    put:
      description: Set a connector as the default for the organization
      parameters:
      - description: Connector ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Set default connector
      tags:
      - connectors
  /connectors/{id}/schema:
    get:
      description: Get the database schema for a connector
      parameters:
      - description: Connector ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get connector schema
      tags:
      - connectors
  /connectors/{id}/test:
    post:
      description: Test connection to a database connector
      parameters:
      - description: Connector ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Test a connector
      tags:
      - connectors
  /connectors/test:
    post:
      consumes:
      - application/json
      description: Test a database connection using raw config before saving
      parameters:
      - description: Connector details to test
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Test a connector configuration
      tags:
      - connectors
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"